#50: Refactor to rd_process.
