+++ src/raindrop_todoist_syncer/main.py
@@ -2,13 +2,13 @@ import datetime
 import traceback
 from loguru import logger
 
+from raindrop_todoist_syncer.rd_process import RaindropsProcessor
 from raindrop_todoist_syncer.raindrop import (
     DatabaseManager,
     EnvironmentVariablesFileManager,
     RaindropAccessTokenRefresher,
     RaindropClient,
     RaindropCredentialsManager,
-    RaindropsProcessor,
 )
 from raindrop_todoist_syncer.todoist import TodoistTaskCreator
